backend/banking-rest-api-develop-using-python-django/banksystem/requirements.txt
Django == 3.0.6
djangorestframework==3.12.4
Markdown==3.3.6
django-filter==21.1
djongo
pymongo==3.12.3
sqlparse==0.2.4
django-cors-headers==3.10.1
#setuptools
